move subnets to their own file
