+++ beacon-chain/p2p/BUILD.bazel
@@ -20,6 +20,7 @@ go_library(
         "rpc_topic_mappings.go",
         "sender.go",
         "service.go",
+        "subnets.go",
         "utils.go",
         "watch_peers.go",
     ],
@@ -89,6 +90,7 @@ go_test(
         "parameter_test.go",
         "sender_test.go",
         "service_test.go",
+        "subnets_test.go",
     ],
     embed = [":go_default_library"],
     flaky = True,
